moved sensorfusion to its own class
